requirements.txt
asgiref==3.6.0
aspose-words==23.3.0
aspose-words-cloud==23.2.0
Babel==2.12.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.6
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
